+++ .stylelintrc.json
@@ -2,7 +2,9 @@
   "extends": ["stylelint-config-standard"],
   "plugins": ["stylelint-order"],
   "rules": {
-    "order/properties-alphabetical-order": true
+    "order/properties-alphabetical-order": true,
+    "selector-class-pattern": null,
+    "block-no-empty": null
   },
   "ignoreFiles": "node_modules/**/*"
 }
