+++ src/Components/Progress/Progress.module.scss
@@ -1,0 +1,12 @@
+.stepperBox {
+  background: #f3f3f3;
+  border-radius: 15px;
+  margin: 10px auto;
+  padding: 10px;
+  width: 90%;
+}
+
+.successPost {
+  display: flex;
+  flex-direction: column;
+}
